Remove all mkv-vesting duplicate code
